% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifski.R
\name{gifski}
\alias{gifski}
\title{Gifski}
\usage{
gifski(png_files, gif_file = "animation.gif", width = 480, height = 480,
  delay = 1, loop = TRUE, progress = TRUE)
}
\arguments{
\item{png_files}{vector of png files}

\item{gif_file}{output gif file}

\item{width}{gif width in pixels}

\item{height}{gif height in pixel}

\item{delay}{time to show each image in seconds}

\item{loop}{should the gif play forever (FALSE to only play once)}

\item{progress}{show progress bar}
}
\description{
Gifski converts video frames to GIF animations using pngquant's fancy features
for efficient cross-frame palettes and temporal dithering. It produces animated
GIFs that use thousands of colors per frame.
}
\examples{
png("frame\%03d.png")
par(ask = FALSE)
for(i in 1:10)
  plot(rnorm(i * 10), main = i)
dev.off()
png_files <- sprintf("frame\%03d.png", 1:10)
gif_file <- gifski(png_files)
unlink(png_files)
utils::browseURL(gif_file)
}
