% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{streaming}
\alias{streaming}
\alias{gifski_encoder_init}
\title{Streaming Gifski Encoder}
\usage{
gifski_encoder_init(gif_file = "animation.gif", width = 800,
  height = 600, loop = TRUE, delay = 1)
}
\arguments{
\item{gif_file}{output gif file}

\item{width}{gif width in pixels}

\item{height}{gif height in pixel}

\item{loop}{should the gif play forever (FALSE to only play once)}

\item{delay}{time to show each image in seconds}
}
\description{
Starts a stateful gif encoder and returns a closure to add png frames
to the gif. Upon completion, the caller has to invoke the closure one
more time with `png_file = NULL` to finalize the output.
}
\seealso{
Other gifski: \code{\link{gifski}}
}
\concept{gifski}
